refactor(storage): modularize storage subsystem and rename read_prefix to submit_prefix
Split monolithic storage implementation into focused modules:
- block_entry.rs: BlockEntry with PinToken reference counting
- insert_worker.rs: background insertion and GC logic
- prefetch.rs: prefix-based prefetching orchestration
- sealed_block.rs: immutable block representation

Key changes:
- Rename BackingStore::read_prefix to submit_prefix for clarity
- Add PinToken for explicit block pinning lifecycle management
- Export PinToken from public API
- Remove unused dependencies (colored, parking_lot, serde_json, logforth)
